Create Problemstatement
The purpose of the requirements document is to systematically capture requirements for the project and the system “Member Registration Portal” to be developed. The application should be Cloud Native Architecture with Microservices. Both functional and non-functional requirements are captured in this document. It also serves as the input for the project scoping.
About the System
The client would like to develop an independent application Member Registration Portal (MRP) application in order to automate the process of registering a member, edit his details and submit his claims
Scope of the System
The scope of the system is explained through its modules as follows
⦁	Member Registration – used by member to register the details of self-information into the system. The system stores the details of the member in the system and able to edit it.
⦁	Submit claims – The user should be able to submit the claim for him and his dependent
⦁	Architecture Diagram for the Problem Statement
Use case Diagram